speech/requirements.txt
pydub
youtube_dl
ffmpeg-python
pandas
google-cloud-speech
google-cloud-storage
google-auth
googleapis-common-protos

